update(internal_clients): sessionrequests work with external authn
