+++ internal_clients/docker-compose.dev.yml
@@ -5,8 +5,6 @@ version: "3"
 services:
   webapi:
     container_name: internal_clients_webapi
-    extra_hosts:
-      - "authn.briantaylorvann.com:10.88.0.1"
     build:
       context: ./
       dockerfile: dockerfile
